code/requirements.txt
requests
boto3
dnspython